+++ smokclient/extras/__init__.py
@@ -1,0 +1,3 @@
+from .pp_database import BasePathpointDatabase, BaseDataToSynchronize
+
+__all__ = ['BasePathpointDatabase', 'BaseDataToSynchronize']