+++ smokclient/extras/pp_database/__init__.py
@@ -1,0 +1,3 @@
+from .base import BasePathpointDatabase, BaseDataToSynchronize
+
+__all__ = ['BasePathpointDatabase', 'BaseDataToSynchronize']