+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v0.0.3
+
+* added pluggable pathpoint value databases
+